+++ packages/evoke-charts/src/index.js
@@ -1,6 +1,7 @@
 import EvChart from './chart.vue'
 import { useChart, provideChartContext, injectChartContext } from './useChart'
 import { applySeriesPalette, clearSeriesPalette } from './palette'
+import { chartOptionsSchema, validateOptions } from './schema'
 
 const components = {
   EvChart,
@@ -25,6 +26,9 @@ export {
   // 配色方案
   applySeriesPalette,
   clearSeriesPalette,
+  // Spec 契约
+  chartOptionsSchema,
+  validateOptions,
   // Install
   install,
   components,